+++ src/support/utils.js
@@ -14,7 +14,7 @@ import { context as h2, h1 } from '@adobe/fetch';
 import { hasText, resolveCustomerSecretsName } from '@adobe/spacecat-shared-utils';
 import URI from 'urijs';
 import { JSDOM } from 'jsdom';
-import { loadSecrets } from '@adobe/helix-shared-secrets';
+import { GetSecretValueCommand, SecretsManagerClient } from '@aws-sdk/client-secrets-manager';
 
 URI.preventInvalidHostname = true;
 
@@ -149,16 +149,21 @@ export function getUrlWithoutPath(url) {
  *
  * @param {string} baseURL - The base URL for which the RUM domain key is to be retrieved.
  * @param {UniversalContext} context - Helix Universal Context. See https://github.com/adobe/helix-universal/blob/main/src/adapter.d.ts#L120
- * @param {Object} [opts={}] - Optional parameters for loading secrets. See: https://github.com/adobe/helix-shared/blob/main/packages/helix-shared-secrets/src/secrets-wrapper.d.ts#L18
  * @returns {Promise<string>} - A promise that resolves to the RUM domain key.
  * @throws {Error} Throws an error if no domain key is found for the specified base URL.
  */
-export async function getRUMDomainkey(baseURL, context, opts = {}) {
+export async function getRUMDomainkey(baseURL, context) {
   const customerSecretName = resolveCustomerSecretsName(baseURL, context);
-  const secrets = await loadSecrets(context, { ...opts, name: customerSecretName });
-  if (!hasText(secrets?.RUM_BUNDLER_DOMAIN_KEY)) {
-    throw new Error(`No domainkey found for ${baseURL}`);
+  const { runtime } = context;
+
+  try {
+    const client = new SecretsManagerClient({ region: runtime.region });
+    const command = new GetSecretValueCommand({
+      SecretId: customerSecretName,
+    });
+    const response = await client.send(command);
+    return JSON.parse(response.SecretString)?.RUM_DOMAIN_KEY;
+  } catch (error) {
+    throw new Error(`Error retrieving the domain key for ${baseURL}. Error: ${error.message}`);
   }
-  /* c8 ignore next */
-  return secrets.RUM_BUNDLER_DOMAIN_KEY;
 }